Button and help text layout changes in access rules section
