+++ webapp/channels/src/components/admin_console/access_control/editors/shared.scss
@@ -53,19 +53,3 @@
         opacity: 0.4;
     }
 }
-
-.editor__help-text {
-    color: rgba(var(--center-channel-color-rgb), 0.6);
-    font-size: 12px;
-    line-height: 1.6;
-    
-    a {
-        display: inline-block;
-        margin-top: 8px;
-        color: var(--link-color);
-        
-        &:hover {
-            text-decoration: underline;
-        }
-    }
-}
