+++ webapp/channels/src/components/admin_console/access_control/editors/shared.tsx
@@ -4,8 +4,6 @@
 import React from 'react';
 import {FormattedMessage} from 'react-intl';
 
-import Markdown from 'components/markdown';
-
 import './shared.scss';
 
 interface TestButtonProps {
@@ -18,15 +16,10 @@ interface AddAttributeButtonProps {
     disabled: boolean;
 }
 
-interface HelpTextProps {
-    message: string;
-    onLearnMoreClick: () => void;
-}
-
 export function TestButton({onClick, disabled}: TestButtonProps): JSX.Element {
     return (
         <button
-            className='editor__test-btn'
+            className='btn btn-sm btn-tertiary'
             onClick={onClick}
             disabled={disabled}
         >
@@ -42,7 +35,7 @@ export function TestButton({onClick, disabled}: TestButtonProps): JSX.Element {
 export function AddAttributeButton({onClick, disabled}: AddAttributeButtonProps): JSX.Element {
     return (
         <button
-            className='editor__add-row-button'
+            className='btn btn-sm btn-tertiary'
             onClick={onClick}
             disabled={disabled}
         >
@@ -52,26 +45,5 @@ export function AddAttributeButton({onClick, disabled}: AddAttributeButtonProps)
                 defaultMessage='Add attribute'
             />
         </button>
-    );
-}
-
-export function HelpText({message, onLearnMoreClick}: HelpTextProps): JSX.Element {
-    return (
-        <div className='editor__help-text'>
-            <Markdown
-                message={message}
-                options={{mentionHighlight: false}}
-            />
-            <a
-                href='#'
-                className='editor__learn-more'
-                onClick={onLearnMoreClick}
-            >
-                <FormattedMessage
-                    id='admin.access_control.cel.learnMore'
-                    defaultMessage='Learn more about creating access expressions with examples.'
-                />
-            </a>
-        </div>
     );
 }
